+++ src/tgl/d3drm/impl.h
@@ -1,5 +1,4 @@
 
-#include "utils.h"
 #include "tgl/tgl.h"
 
 #include <d3drm.h>
@@ -55,10 +54,10 @@ namespace TglImpl {
 		View* CreateView(
 			const Device*,
 			const Camera*,
-			unsigned long x,
-			unsigned long y,
-			unsigned long width,
-			unsigned long height
+			unsigned int x,
+			unsigned int y,
+			unsigned int width,
+			unsigned int height
 		) override;
 		Camera* CreateCamera() override;
 		Light* CreateLight(LightType, float r, float g, float b) override;
@@ -77,10 +76,10 @@ namespace TglImpl {
 		) override;
 		Texture* CreateTexture() override;
 
-		Result SetTextureDefaultShadeCount(unsigned long) override;
+		Result SetTextureDefaultShadeCount(unsigned int) override;
 
 		// vtable+0x30
-		Result SetTextureDefaultColorCount(unsigned long) override;
+		Result SetTextureDefaultColorCount(unsigned int) override;
 
 		HRESULT CreateTextureFromSurface(LPDIRECTDRAWSURFACE pSurface, LPDIRECT3DRMTEXTURE2* pTexture2) {
 			return m_data->CreateTextureFromSurface(pSurface, pTexture2);
@@ -127,13 +126,13 @@ namespace TglImpl {
 		void* ImplementationDataPtr() override;
 
 		// vtable+0x08
-		unsigned long GetWidth() override;
-		unsigned long GetHeight() override;
+		unsigned int GetWidth() override;
+		unsigned int GetHeight() override;
 
 		// vtable+0x10
 		Result SetColorModel(ColorModel) override;
 		Result SetShadingModel(ShadingModel) override;
-		Result SetShadeCount(unsigned long) override;
+		Result SetShadeCount(unsigned int) override;
 		Result SetDither(int) override;
 
 		// vtable+0x20
@@ -177,14 +176,14 @@ namespace TglImpl {
 		Result GetBackgroundColor(float* r, float* g, float* b) override;
 		Result Clear() override;
 		Result Render(const Group*) override;
-		Result ForceUpdate(unsigned long x, unsigned long y, unsigned long width, unsigned long height) override;
+		Result ForceUpdate(unsigned int x, unsigned int y, unsigned int width, unsigned int height) override;
 
 		// vtable+0x30
 		Result TransformWorldToScreen(const float world[3], float screen[4]) override;
 		Result TransformScreenToWorld(const float screen[4], float world[3]) override;
 		Result Pick(
-			unsigned long x,
-			unsigned long y,
+			unsigned int x,
+			unsigned int y,
 			const Group** ppGroupsToPickFrom,
 			int groupsToPickFromCount,
 			const Group**& rppPickedGroups,
@@ -348,13 +347,13 @@ namespace TglImpl {
 
 		// vtable+0x08
 		Mesh* CreateMesh(
-			unsigned long faceCount,
-			unsigned long vertexCount,
+			unsigned int faceCount,
+			unsigned int vertexCount,
 			float (*pPositions)[3],
 			float (*pNormals)[3],
 			float (*pTextureCoordinates)[2],
-			unsigned long (*pFaceIndices)[3],
-			unsigned long (*pTextureIndices)[3],
+			unsigned int (*pFaceIndices)[3],
+			unsigned int (*pTextureIndices)[3],
 			ShadingModel shadingModel
 		) override;
 		Result GetBoundingBox(float min[3], float max[3]) const override;
@@ -369,13 +368,13 @@ namespace TglImpl {
 	private:
 		inline Result CreateMeshImpl(
 			MeshImpl* pMeshImpl,
-			unsigned long faceCount,
-			unsigned long vertexCount,
+			unsigned int faceCount,
+			unsigned int vertexCount,
 			float (*pPositions)[3],
 			float (*pNormals)[3],
 			float (*pTextureCoordinates)[2],
-			unsigned long (*pFaceIndices)[3],
-			unsigned long (*pTextureIndices)[3],
+			unsigned int (*pFaceIndices)[3],
+			unsigned int (*pTextureIndices)[3],
 			ShadingModel shadingModel
 		);
